requirements.txt
openai==1.0.0
mysql-connector-python==8.0.33
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
